# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TontineChain API",
        "description": "Documentation de l'API TontineChain pour le Hackathon MIABE 2026.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Local"
        },
        {
            "url": "https://tonnine-benin-backend.onrender.com",
            "description": "Serveur de Production"
        }
    ],
    "paths": {
...
        "/api/v1/groups": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Lister mes groupes",
                "operationId": "8731199f9d1e321702524c8df8c5834c",
                "responses": {
                    "200": {
                        "description": "Liste des groupes"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Créer une nouvelle tontine",
                "operationId": "0dbd5970197eb93f2a1ed934c341aeab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tontine Famille"
                                    },
                                    "contribution_amount": {
                                        "type": "number",
                                        "example": 10000
                                    },
                                    "max_members": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "enum": [
                                            "weekly",
                                            "biweekly",
                                            "monthly"
                                        ]
                                    },
                                    "payout_method": {
                                        "type": "string",
                                        "enum": [
                                            "sequential",
                                            "random",
                                            "bidding"
                                        ]
                                    },
                                    "insurance_opt_in": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Groupe créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Détails d'un groupe",
                "operationId": "cc64455d914e608b0b7f53d3b73b59bb",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du groupe"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/start": {
            "post": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Démarrer la tontine",
                "operationId": "4716379a3472bb4db88043afb5f609dd",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tontine démarrée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "contract_address": {
                                            "type": "string"
                                        },
                                        "demo_notice": {
                                            "properties": {
                                                "is_simulation": {
                                                    "type": "boolean"
                                                },
                                                "blockchain_tx": {
                                                    "type": "string"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/invite": {
            "post": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Inviter un membre",
                "operationId": "5a69ca01630f7b36a7e1f59f01f9694c",
                "responses": {
                    "200": {
                        "description": "Invitation envoyée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "member_analysis": {
                                            "type": "object"
                                        },
                                        "demo_notice": {
                                            "properties": {
                                                "is_simulation": {
                                                    "type": "boolean"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/groups/{group}/stats": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Statistiques & Analyse de Risque IA",
                "operationId": "3e7d5fc393168a14146671e8fc31e1ad",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stats et prédiction de risque"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/contract": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Télécharger le contrat PDF de la tontine",
                "operationId": "903230a31f2478c27c456d8db122ff16",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier PDF du contrat"
                    },
                    "404": {
                        "description": "Contrat non disponible (tontine non démarrée)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Utiliser le token Bearer retourné par /api/v1/auth/verify-otp",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}